odahu-tools batch log expects to path to model but -model and -version strings
